Add Hawaiian 'okina to Nimbus family fonts
When using the Flattery theme on an iPod 4G, Hawaiian music did not
display the proper character for the Hawaiian 'okina in song metadata.

This commit adds the Hawaiian 'okina to Nimbus family of fonts by
copying the bitmap of the `quoteleft` to a new `okina`.

This commit also happens to include the same change made to the
`14-Nimbus-padded.fnt` that the Flattery theme uses.